# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ConflictCheckRequest": {
        "properties": {
          "party_names": {
            "items": {
              "type": "string"
            },
            "title": "Party Names",
            "type": "array"
          }
        },
        "required": [
          "party_names"
        ],
        "title": "ConflictCheckRequest",
        "type": "object"
      },
      "ConflictCheckResponse": {
        "properties": {
          "conflicts": {
            "items": {
              "$ref": "#/components/schemas/ConflictResult"
            },
            "title": "Conflicts",
            "type": "array"
          },
          "has_conflicts": {
            "title": "Has Conflicts",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "has_conflicts",
          "conflicts"
        ],
        "title": "ConflictCheckResponse",
        "type": "object"
      },
      "ConflictResult": {
        "properties": {
          "matched_name": {
            "title": "Matched Name",
            "type": "string"
          },
          "matter_id": {
            "title": "Matter Id",
            "type": "string"
          },
          "matter_title": {
            "title": "Matter Title",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "searched_name": {
            "title": "Searched Name",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "searched_name",
          "matched_name",
          "role",
          "matter_id",
          "matter_title",
          "status"
        ],
        "title": "ConflictResult",
        "type": "object"
      },
...
      "DocumentResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "failure_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Reason"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "latest_revision_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Revision Id"
          },
          "matter_id": {
            "title": "Matter Id",
            "type": "string"
          },
          "provenance_state": {
            "title": "Provenance State",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "matter_id",
          "title",
          "status",
          "provenance_state",
          "created_at"
        ],
        "title": "DocumentResponse",
        "type": "object"
      },
...
      "DraftSummaryResponse": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "matter_id": {
            "title": "Matter Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "matter_id",
          "title",
          "version",
          "status",
          "updated_at"
        ],
        "title": "DraftSummaryResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MatterCreate": {
        "properties": {
          "ai_processing_policy": {
            "default": "standard",
            "description": "AI processing policy",
            "title": "Ai Processing Policy",
            "type": "string"
          },
          "case_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Type of case",
            "title": "Case Type"
          },
          "client_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Client name",
            "title": "Client Name"
          },
          "confidentiality_level": {
            "default": "standard",
            "description": "Confidentiality level",
            "title": "Confidentiality Level",
            "type": "string"
          },
          "internal_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Internal reference ID",
            "title": "Internal Reference"
          },
          "jurisdiction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "The jurisdiction of the matter",
            "title": "Jurisdiction"
          },
          "title": {
            "description": "The title of the matter",
            "maxLength": 255,
            "minLength": 3,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title"
        ],
        "title": "MatterCreate",
        "type": "object"
      },
      "MatterPartyResponse": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "role",
          "type"
        ],
        "title": "MatterPartyResponse",
        "type": "object"
      },
      "MatterResponse": {
        "properties": {
          "access_scope": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Access Scope"
          },
          "ai_processing_policy": {
            "title": "Ai Processing Policy",
            "type": "string"
          },
          "case_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Type"
          },
          "client_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Name"
          },
          "closed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed At"
          },
          "confidentiality_level": {
            "title": "Confidentiality Level",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "internal_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Internal Reference"
          },
          "jurisdiction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jurisdiction"
          },
          "opened_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opened At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "status",
          "confidentiality_level",
          "ai_processing_policy"
        ],
        "title": "MatterResponse",
        "type": "object"
      },
...
      "OverrideConflictRequest": {
        "properties": {
          "reason": {
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "OverrideConflictRequest",
        "type": "object"
      },
      "OverrideConflictResponse": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "message"
        ],
        "title": "OverrideConflictResponse",
        "type": "object"
      },
...
      "QAResponse": {
        "additionalProperties": false,
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/QACitation"
            },
            "title": "Citations",
            "type": "array"
          },
          "degraded_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Degraded Reason"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "answer",
          "status"
        ],
        "title": "QAResponse",
        "type": "object"
      },
      "QuestionRequest": {
        "additionalProperties": false,
        "properties": {
          "question": {
            "maxLength": 4096,
            "minLength": 1,
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "question"
        ],
        "title": "QuestionRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "MESA Law API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/documents/matters/{matter_id}": {
      "get": {
        "operationId": "listMatterDocuments",
        "parameters": [
          {
            "in": "path",
            "name": "matter_id",
            "required": true,
            "schema": {
              "title": "Matter Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DocumentResponse"
                  },
                  "title": "Response Listmatterdocuments",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Matter Documents"
      }
    },
...
    "/api/v1/draft-studio/drafts/matter/{matter_id}": {
      "get": {
        "operationId": "listMatterDrafts",
        "parameters": [
          {
            "in": "path",
            "name": "matter_id",
            "required": true,
            "schema": {
              "title": "Matter Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DraftSummaryResponse"
                  },
                  "title": "Response Listmatterdrafts",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Matter Drafts",
        "tags": [
          "draft-studio"
        ]
      }
    },
...
    "/api/v1/matters": {
      "get": {
        "operationId": "listMatters",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MatterResponse"
                  },
                  "title": "Response Listmatters",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Matters"
      },
      "post": {
        "operationId": "createMatter",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatterCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatterResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Matter"
      }
    },
    "/api/v1/matters/conflict-check": {
      "post": {
        "operationId": "conflictCheck",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConflictCheckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictCheckResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Check Conflicts"
      }
    },
    "/api/v1/matters/{matter_id}": {
      "get": {
        "operationId": "getMatter",
        "parameters": [
          {
            "in": "path",
            "name": "matter_id",
            "required": true,
            "schema": {
              "title": "Matter Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatterResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Matter"
      }
    },
...
    "/api/v1/matters/{matter_id}/override-conflict": {
      "post": {
        "operationId": "overrideConflict",
        "parameters": [
          {
            "in": "path",
            "name": "matter_id",
            "required": true,
            "schema": {
              "title": "Matter Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "max_age_seconds",
            "required": false,
            "schema": {
              "default": 300,
              "title": "Max Age Seconds",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OverrideConflictRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OverrideConflictResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Override Conflict"
      }
    },
    "/api/v1/matters/{matter_id}/parties": {
      "get": {
        "operationId": "listMatterParties",
        "parameters": [
          {
            "in": "path",
            "name": "matter_id",
            "required": true,
            "schema": {
              "title": "Matter Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MatterPartyResponse"
                  },
                  "title": "Response Listmatterparties",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Matter Parties"
      }
    },
    "/api/v1/matters/{matter_id}/qa": {
      "post": {
        "operationId": "matterQA",
        "parameters": [
          {
            "in": "path",
            "name": "matter_id",
            "required": true,
            "schema": {
              "title": "Matter Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QAResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Matter Qa Endpoint"
      }
    },
    "/api/v1/matters/{matter_id}/rebuild-mesa": {
      "post": {
        "operationId": "rebuildMatterMesa",
        "parameters": [
          {
            "in": "path",
            "name": "matter_id",
            "required": true,
            "schema": {
              "title": "Matter Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rebuild Matter Mesa"
      }
    },
...
  }
}